Change class names
