+++ java/memory/src/main/java/org/apache/arrow/memory/NettyAllocationManager.java
@@ -1,0 +1,64 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.arrow.memory;
+
+import io.netty.buffer.PooledByteBufAllocatorL;
+import io.netty.buffer.UnsafeDirectLittleEndian;
+
+/**
+ * The default implementation of AllocationManagerBase. The implementation is responsible for managing when memory
+ * is allocated and returned to the Netty-based PooledByteBufAllocatorL.
+ */
+public class NettyAllocationManager extends AllocationManager {
+
+  private static final PooledByteBufAllocatorL INNER_ALLOCATOR = new PooledByteBufAllocatorL();
+  static final UnsafeDirectLittleEndian EMPTY = INNER_ALLOCATOR.empty;
+  static final long CHUNK_SIZE = INNER_ALLOCATOR.getChunkSize();
+
+  private final int size;
+  private final UnsafeDirectLittleEndian memoryChunk;
+
+  NettyAllocationManager(BaseAllocator accountingAllocator, int size) {
+    super(accountingAllocator);
+    this.memoryChunk = INNER_ALLOCATOR.allocate(size);
+    this.size = memoryChunk.capacity();
+  }
+
+  /**
+   * Get the underlying memory chunk managed by this AllocationManager.
+   * @return buffer
+   */
+  UnsafeDirectLittleEndian getMemoryChunk() {
+    return memoryChunk;
+  }
+
+  @Override
+  long memoryAddress() {
+    return memoryChunk.memoryAddress();
+  }
+
+  @Override
+  protected void release0() {
+    memoryChunk.release();
+  }
+
+  @Override
+  protected int getSize() {
+    return size;
+  }
+}
